Move unit tests down into sage app sdk
