--- app/src/androidTest/resources/test_scheduled_activities.json
@@ -1,197 +1,0 @@
-{
-  "items": [
-    {
-      "guid": "273c4518-7cb6-4496-b1dd-c0b5bf291b09:2018-08-17T00:00:00.000",
-      "schedulePlanGuid": "e24e6601-1822-48b0-8770-00870d870708",
-      "activity": {
-        "label": "Medication",
-        "guid": "273c4518-7cb6-4496-b1dd-c0b5bf291b09",
-        "task": {
-          "identifier": "Medication",
-          "type": "TaskReference"
-        },
-        "activityType": "task",
-        "type": "Activity"
-      },
-      "persistent": false,
-      "scheduledOn": "2018-08-17T00:00:00.000-04:00",
-      "expiresOn": "2018-08-18T00:00:00.000-04:00",
-      "status": "expired",
-      "type": "ScheduledActivity"
-    },
-    {
-      "guid": "fe79d987-28a2-4ccd-bcf3-b3d07b925a6b:2018-08-17T00:00:00.000",
-      "schedulePlanGuid": "9ac67c85-e8fb-4f4c-a4ef-f71aba7cd0b2",
-      "activity": {
-        "label": "Study Burst Task",
-        "guid": "fe79d987-28a2-4ccd-bcf3-b3d07b925a6b",
-        "task": {
-          "identifier": "study-burst-task",
-          "type": "TaskReference"
-        },
-        "activityType": "task",
-        "type": "Activity"
-      },
-      "persistent": false,
-      "scheduledOn": "2018-08-17T00:00:00.000-04:00",
-      "expiresOn": "2018-08-18T00:00:00.000-04:00",
-      "status": "expired",
-      "type": "ScheduledActivity"
-    },
-    {
-      "guid": "a341c893-615d-48e1-ab6a-d418af720269:2018-08-17T13:40:39.183",
-      "schedulePlanGuid": "e4c528e6-82f3-4726-9a46-c6dc1f89c366",
-      "startedOn": "2018-08-17T17:44:05.114Z",
-      "finishedOn": "2018-08-17T18:09:28.610Z",
-      "activity": {
-        "label": "Motivation",
-        "guid": "a341c893-615d-48e1-ab6a-d418af720269",
-        "survey": {
-          "identifier": "Motivation",
-          "guid": "7beb8a71-65d0-4217-a4fb-3bb029b55258",
-          "createdOn": "2018-08-07T22:41:52.186Z",
-          "href": "https://ws.sagebridge.org/v3/surveys/7beb8a71-65d0-4217-a4fb-3bb029b55258/revisions/2018-08-07T22:41:52.186Z",
-          "type": "SurveyReference"
-        },
-        "activityType": "survey",
-        "type": "Activity"
-      },
-      "persistent": false,
-      "clientData": {
-        "motivation_II_T": "eat"
-      },
-      "scheduledOn": "2018-08-17T13:40:39.183-04:00",
-      "status": "finished",
-      "type": "ScheduledActivity"
-    },
-    {
-      "guid": "178ef89c-78b9-4861-b308-3dda0daf756d:2018-08-17T13:40:39.183",
-      "schedulePlanGuid": "d5479fc2-2d42-4db3-83d9-6cca6558af2e",
-      "activity": {
-        "label": "Health Survey",
-        "labelDetail": "4 Minutes",
-        "guid": "178ef89c-78b9-4861-b308-3dda0daf756d",
-        "survey": {
-          "identifier": "Demographics",
-          "guid": "f2637b2a-b473-4f84-b645-1431a3b4418e",
-          "createdOn": "2018-07-26T00:40:57.836Z",
-          "href": "https://ws.sagebridge.org/v3/surveys/f2637b2a-b473-4f84-b645-1431a3b4418e/revisions/2018-07-26T00:40:57.836Z",
-          "type": "SurveyReference"
-        },
-        "activityType": "survey",
-        "type": "Activity"
-      },
-      "persistent": false,
-      "scheduledOn": "2018-08-17T13:40:39.183-04:00",
-      "status": "available",
-      "type": "ScheduledActivity"
-    },
-    {
-      "guid": "e6fe761f-6187-4e8f-b659-bce4edc98f06:2018-08-17T13:40:39.183",
-      "schedulePlanGuid": "c4ec16b4-2985-4ae3-a17a-936ec52a4218",
-      "activity": {
-        "label": "Engagement Survey",
-        "labelDetail": "8 minutes",
-        "guid": "e6fe761f-6187-4e8f-b659-bce4edc98f06",
-        "survey": {
-          "identifier": "Engagement",
-          "guid": "56fbeb25-0a90-499a-a0d8-3c4935b60312",
-          "createdOn": "2018-07-23T22:42:34.300Z",
-          "href": "https://ws.sagebridge.org/v3/surveys/56fbeb25-0a90-499a-a0d8-3c4935b60312/revisions/2018-07-23T22:42:34.300Z",
-          "type": "SurveyReference"
-        },
-        "activityType": "survey",
-        "type": "Activity"
-      },
-      "persistent": false,
-      "scheduledOn": "2018-08-17T13:40:39.183-04:00",
-      "status": "available",
-      "type": "ScheduledActivity"
-    },
-    {
-      "guid": "60868b71-30a4-4e04-a00b-3aca6651deb2:2018-08-17T13:40:39.183",
-      "schedulePlanGuid": "bdd89e43-819a-4bba-8ce9-421bd09518a2",
-      "activity": {
-        "label": "Symptom",
-        "guid": "60868b71-30a4-4e04-a00b-3aca6651deb2",
-        "task": {
-          "identifier": "Symptoms",
-          "type": "TaskReference"
-        },
-        "activityType": "task",
-        "type": "Activity"
-      },
-      "persistent": true,
-      "scheduledOn": "2018-08-18T13:40:39.183-04:00",
-      "status": "available",
-      "type": "ScheduledActivity"
-    },
-    {
-      "guid": "2dc5dddb-cdb8-4291-9e98-183dd9b20e5f:2016-03-08T17:24:02.103",
-      "schedulePlanGuid": "0ef97f95-9b47-445a-b7c3-99864aab1454",
-      "activity": {
-        "label": "Compound Activity From Definition",
-        "guid": "2dc5dddb-cdb8-4291-9e98-183dd9b20e5f",
-        "compoundActivity": {
-          "schemaList": [
-            {
-              "id": "simple-test-schema",
-              "revision": 2,
-              "type": "SchemaReference"
-            }
-          ],
-          "surveyList": [
-            {
-              "identifier": "simple-survey-1",
-              "guid": "2a93e632-c760-4b52-b567-ec36b26acfa4",
-              "createdOn": "2017-02-09T02:26:14.644Z",
-              "href": "http://localhost:9000/v3/surveys/2a93e632-c760-4b52-b567-ec36b26acfa4/revisions/2017-02-09T02:26:14.644Z",
-              "type": "SurveyReference"
-            }
-          ],
-          "taskIdentifier": "compound-from-def",
-          "type": "CompoundActivity"
-        },
-        "activityType": "compound",
-        "type": "Activity"
-      },
-      "persistent": false,
-      "scheduledOn": "2016-03-08T17:24:02.103-04:00",
-      "expiresOn": "2016-03-09T17:24:02.103-04:00",
-      "status": "expired",
-      "type": "ScheduledActivity"
-    },
-    {
-      "guid": "f34e5be9-df81-46a1-bf21-5a2994c44e9d:2016-03-08T17:24:02.103",
-      "schedulePlanGuid": "0ef97f95-9b47-445a-b7c3-99864aab1454",
-      "activity": {
-        "label": "Task Activity with Schema",
-        "guid": "f34e5be9-df81-46a1-bf21-5a2994c44e9d",
-        "task": {
-          "identifier": "back-compat-task",
-          "schema": {
-            "id": "upload-v2-manual-test",
-            "revision": 1,
-            "type": "SchemaReference"
-          },
-          "type": "TaskReference"
-        },
-        "activityType": "task",
-        "type": "Activity"
-      },
-      "persistent": false,
-      "scheduledOn": "2016-03-08T17:24:02.103-04:00",
-      "expiresOn": "2016-03-09T17:24:02.103-04:00",
-      "status": "expired",
-      "type": "ScheduledActivity"
-    }],
-  "requestParams": {
-    "startTime": "2018-08-21T00:00:00.000-04:00",
-    "endTime": "2018-08-24T00:00:00.000-04:00",
-    "type": "RequestParams"
-  },
-  "startTime": "2018-08-21T00:00:00.000-04:00",
-  "endTime": "2018-08-24T00:00:00.000-04:00",
-  "total": 2,
-  "type": "DateTimeRangeResourceList"
-}
