--- app/src/androidTest/resources/test_schedules_finished_between.json
@@ -1,102 +1,0 @@
-{
-  "items": [
-    {
-      "guid": "273c4518-7cb6-4496-b1dd-c0b5bf291b09:2018-08-17T00:00:00.000",
-      "schedulePlanGuid": "e24e6601-1822-48b0-8770-00870d870708",
-      "startedOn": "2018-08-17T17:44:05.114Z",
-      "finishedOn": "2018-08-17T18:09:28.610Z",
-      "activity": {
-        "label": "Medication",
-        "guid": "273c4518-7cb6-4496-b1dd-c0b5bf291b09",
-        "task": {
-          "identifier": "Medication",
-          "type": "TaskReference"
-        },
-        "activityType": "task",
-        "type": "Activity"
-      },
-      "persistent": false,
-      "clientData": {
-        "motivation_II_T": "eat"
-      },
-      "scheduledOn": "2018-08-17T00:00:00.000-04:00",
-      "expiresOn": "2018-08-18T00:00:00.000-04:00",
-      "status": "expired",
-      "type": "ScheduledActivity"
-    },
-    {
-      "guid": "a341c893-615d-48e1-ab6a-d418af720269:2018-08-17T00:00:00.000-04:00",
-      "schedulePlanGuid": "e4c528e6-82f3-4726-9a46-c6dc1f89c366",
-      "startedOn": "2018-08-17T17:44:05.114Z",
-      "finishedOn": "2018-08-17T18:09:28.610Z",
-      "activity": {
-        "label": "Motivation",
-        "guid": "a341c893-615d-48e1-ab6a-d418af720269",
-        "survey": {
-          "identifier": "Motivation",
-          "guid": "7beb8a71-65d0-4217-a4fb-3bb029b55258",
-          "createdOn": "2018-08-07T22:41:52.186Z",
-          "href": "https://ws.sagebridge.org/v3/surveys/7beb8a71-65d0-4217-a4fb-3bb029b55258/revisions/2018-08-07T22:41:52.186Z",
-          "type": "SurveyReference"
-        },
-        "activityType": "survey",
-        "type": "Activity"
-      },
-      "persistent": false,
-      "clientData": {
-        "motivation_II_T": "eat"
-      },
-      "scheduledOn": "2018-08-17T00:00:00.000-04:00",
-      "status": "finished",
-      "type": "ScheduledActivity"
-    },
-    {
-      "guid": "273c4518-7cb6-4496-b1dd-c0b5bf291b09:2018-08-18T00:00:00.000-04:00",
-      "schedulePlanGuid": "e24e6601-1822-48b0-8770-00870d870708",
-      "activity": {
-        "label": "Medication",
-        "guid": "273c4518-7cb6-4496-b1dd-c0b5bf291b09",
-        "task": {
-          "identifier": "Medication",
-          "type": "TaskReference"
-        },
-        "activityType": "task",
-        "type": "Activity"
-      },
-      "persistent": false,
-      "scheduledOn": "2018-08-18T00:00:00.000-04:00",
-      "expiresOn": "2018-08-19T00:00:00.000-04:00",
-      "status": "expired",
-      "type": "ScheduledActivity"
-    },
-    {
-      "guid": "a341c893-615d-48e1-ab6a-d418af720269:2018-08-18T00:00:00.000-04:00",
-      "schedulePlanGuid": "e4c528e6-82f3-4726-9a46-c6dc1f89c366",
-      "activity": {
-        "label": "Motivation",
-        "guid": "a341c893-615d-48e1-ab6a-d418af720269",
-        "survey": {
-          "identifier": "Motivation",
-          "guid": "7beb8a71-65d0-4217-a4fb-3bb029b55258",
-          "createdOn": "2018-08-07T22:41:52.186Z",
-          "href": "https://ws.sagebridge.org/v3/surveys/7beb8a71-65d0-4217-a4fb-3bb029b55258/revisions/2018-08-07T22:41:52.186Z",
-          "type": "SurveyReference"
-        },
-        "activityType": "survey",
-        "type": "Activity"
-      },
-      "persistent": false,
-      "scheduledOn": "2018-08-18T00:00:00.000-04:00",
-      "status": "finished",
-      "type": "ScheduledActivity"
-    }],
-  "requestParams": {
-    "startTime": "2018-08-21T00:00:00.000-04:00",
-    "endTime": "2018-08-24T00:00:00.000-04:00",
-    "type": "RequestParams"
-  },
-  "startTime": "2018-08-21T00:00:00.000-04:00",
-  "endTime": "2018-08-24T00:00:00.000-04:00",
-  "total": 2,
-  "type": "DateTimeRangeResourceList"
-}
